feat: 사용자 현재 스트릭 조회 API 추가

+++ backend/src/main/java/life/bareun/diary/product/repository/StreakColorRepository.java
@@ -5,7 +5,7 @@ import life.bareun.diary.product.entity.StreakColor;
 import life.bareun.diary.product.entity.StreakColorGrade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StreakColorRepository extends JpaRepository<StreakColor, Long> {
+public interface StreakColorRepository extends JpaRepository<StreakColor, Integer> {
 
     List<StreakColor> findByStreakColorGrade(StreakColorGrade streakColorGrade);
 }
